+++ common/__init__.py
@@ -1,0 +1,5 @@
+"""
+Created on 24 Dec 2020
+
+@author: semuadmin
+"""