+++ pygpsclient/__init__.py
@@ -3,6 +3,7 @@ Created on 27 Sep 2020
 
 @author: semuadmin
 """
+# pylint: disable=invalid-name
 
 from pygpsclient._version import __version__