+++ pygpsclient/settings_frame.py
@@ -7,7 +7,7 @@ Created on 12 Sep 2020
 
 @author: semuadmin
 """
-# pylint: disable=unnecessary-lambda
+# pylint: disable=invalid-name, too-many-instance-attributes, too-many-ancestors, unnecessary-lambda
 
 from tkinter import (
     ttk,
@@ -15,27 +15,19 @@ from tkinter import (
     Button,
     Label,
     Spinbox,
-    Scrollbar,
-    Listbox,
     Scale,
     Checkbutton,
     Radiobutton,
     StringVar,
     IntVar,
-    DoubleVar,
-    N,
-    S,
     E,
     W,
-    LEFT,
     NORMAL,
     DISABLED,
-    VERTICAL,
     HORIZONTAL,
 )
 
 from PIL import ImageTk, Image
-from serial.tools.list_ports import comports
 
 from .globals import (
     ENTCOL,
@@ -46,22 +38,19 @@ from .globals import (
     UMK,
     UI,
     UIK,
-    ADVON,
-    ADVOFF,
     READONLY,
     CONNECTED,
     CONNECTED_FILE,
     DISCONNECTED,
     NOPORTS,
-    KNOWNGPS,
     ICON_CONN,
     ICON_DISCONN,
     ICON_UBXCONFIG,
     ICON_LOGREAD,
-    BAUDRATES,
     NMEA_PROTOCOL,
     UBX_PROTOCOL,
     MIXED_PROTOCOL,
+    PRESELECT,
 )
 from .strings import (
     LBLUBXCONFIG,
@@ -72,6 +61,7 @@ from .strings import (
     LBLSHOWNULL,
     LBLLEGEND,
 )
+from common.serialconfig_frame import SerialConfig
 
 
 class SettingsFrame(Frame):
@@ -90,17 +80,7 @@ class SettingsFrame(Frame):
         self.__master = self.__app.get_master()  # Reference to root class (Tk)
         Frame.__init__(self, self.__master, *args, **kwargs)
 
-        self._show_advanced = False
         self._settings = {}
-        self._ports = ()
-        self._port = StringVar()
-        self._port_desc = StringVar()
-        self._baudrate = IntVar()
-        self._databits = IntVar()
-        self._stopbits = DoubleVar()
-        self._parity = StringVar()
-        self._rtscts = IntVar()
-        self._xonxoff = IntVar()
         self._protocol = IntVar()
         self._raw = IntVar()
         self._autoscroll = IntVar()
@@ -113,7 +93,6 @@ class SettingsFrame(Frame):
         self._record_track = IntVar()
         self._show_zerosig = IntVar()
         self._show_legend = IntVar()
-        self._noports = True
         self._validsettings = True
         self._logpath = None
         self._trackpath = None
@@ -124,7 +103,6 @@ class SettingsFrame(Frame):
 
         self._body()
         self._do_layout()
-        self._get_ports()
         self._reset()
 
     def _body(self):
@@ -138,78 +116,10 @@ class SettingsFrame(Frame):
 
         self.option_add("*Font", self.__app.font_sm)
 
-        # Serial port settings
-        self._frm_basic = Frame(self)
-        self._lbl_port = Label(self._frm_basic, text="Port")
-        self._lbx_port = Listbox(
-            self._frm_basic,
-            border=2,
-            relief="sunken",
-            bg=ENTCOL,
-            width=28,
-            height=5,
-            justify=LEFT,
-            exportselection=False,
-        )
-        self._scr_portv = Scrollbar(self._frm_basic, orient=VERTICAL)
-        self._scr_porth = Scrollbar(self._frm_basic, orient=HORIZONTAL)
-        self._lbx_port.config(yscrollcommand=self._scr_portv.set)
-        self._lbx_port.config(xscrollcommand=self._scr_porth.set)
-        self._scr_portv.config(command=self._lbx_port.yview)
-        self._scr_porth.config(command=self._lbx_port.xview)
-        self._lbx_port.bind("<<ListboxSelect>>", self._on_select_port)
-        self._lbl_baudrate = Label(self._frm_basic, text="Baud rate")
-        self._spn_baudrate = Spinbox(
-            self._frm_basic,
-            values=(BAUDRATES),
-            width=8,
-            state=READONLY,
-            readonlybackground=ENTCOL,
-            wrap=True,
-            textvariable=self._baudrate,
-        )
-        self._btn_toggle = Button(
-            self._frm_basic, text=ADVOFF, width=3, command=self._toggle_advanced
-        )
+        # port configuration settings panel
+        self._frm_serial = SerialConfig(self, PRESELECT)
 
-        self._frm_advanced = Frame(self)
-        self._lbl_databits = Label(self._frm_advanced, text="Data Bits")
-        self._spn_databits = Spinbox(
-            self._frm_advanced,
-            values=(8, 7, 6, 5),
-            width=3,
-            state=READONLY,
-            readonlybackground=ENTCOL,
-            wrap=True,
-            textvariable=self._databits,
-        )
-        self._lbl_stopbits = Label(self._frm_advanced, text="Stop Bits")
-        self._spn_stopbits = Spinbox(
-            self._frm_advanced,
-            values=(2, 1.5, 1),
-            width=3,
-            state=READONLY,
-            readonlybackground=ENTCOL,
-            wrap=True,
-            textvariable=self._stopbits,
-        )
-        self._lbl_parity = Label(self._frm_advanced, text="Parity")
-        self._spn_parity = Spinbox(
-            self._frm_advanced,
-            values=("None", "Even", "Odd", "Mark", "Space"),
-            width=6,
-            state=READONLY,
-            readonlybackground=ENTCOL,
-            wrap=True,
-            textvariable=self._parity,
-        )
-        self._chk_rts = Checkbutton(
-            self._frm_advanced, text="RTS/CTS", variable=self._rtscts
-        )
-        self._chk_xon = Checkbutton(
-            self._frm_advanced, text="Xon/Xoff", variable=self._xonxoff
-        )
-
+        # connection buttons
         self._frm_buttons = Frame(self)
         self._btn_connect = Button(
             self._frm_buttons,
@@ -337,25 +247,9 @@ class SettingsFrame(Frame):
         Position widgets in frame.
         """
 
-        self._frm_basic.grid(column=0, row=0, columnspan=4, sticky=(W, E))
-        self._lbl_port.grid(column=0, row=0, sticky=(W))
-        self._lbx_port.grid(column=1, row=0, sticky=(W, E), padx=3, pady=3)
-        self._scr_portv.grid(column=2, row=0, sticky=(N, S))
-        self._scr_porth.grid(column=1, row=1, sticky=(E, W))
-        self._lbl_baudrate.grid(column=0, row=2, sticky=(W))
-        self._spn_baudrate.grid(column=1, row=2, sticky=(W), padx=3, pady=3)
-        self._btn_toggle.grid(column=2, row=2, sticky=(E))
-
-        self._frm_advanced.grid_forget()
-        self._lbl_databits.grid(column=0, row=0, sticky=(W))
-        self._spn_databits.grid(column=1, row=0, sticky=(W), padx=3, pady=3)
-        self._lbl_stopbits.grid(column=2, row=0, sticky=(W))
-        self._spn_stopbits.grid(column=3, row=0, sticky=(W), padx=3, pady=3)
-        self._lbl_parity.grid(column=0, row=1, sticky=(W))
-        self._spn_parity.grid(column=1, row=1, sticky=(W), padx=3, pady=3)
-        self._chk_rts.grid(column=2, row=1, sticky=(W))
-        self._chk_xon.grid(column=3, row=1, sticky=(W), padx=3, pady=3)
-
+        self._frm_serial.grid(
+            column=0, row=1, columnspan=4, padx=3, pady=3, sticky=(W, E)
+        )
         ttk.Separator(self).grid(
             column=0, row=2, columnspan=4, padx=3, pady=3, sticky=(W, E)
         )
@@ -401,22 +295,6 @@ class SettingsFrame(Frame):
         )
         self._lbl_ubxconfig.grid(column=0, row=9, padx=3, pady=3, sticky=(W))
         self._btn_ubxconfig.grid(column=1, row=9, padx=3, pady=3, sticky=(W))
-
-    def _on_select_port(self, *args, **kwargs):  # pylint: disable=unused-argument
-        """
-        Get selected port from listbox and set global variable.
-        """
-
-        idx = self._lbx_port.curselection()
-        if idx == "":
-            idx = 0
-        port_orig = self._lbx_port.get(idx)
-        port = port_orig[0 : port_orig.find(":")]
-        desc = port_orig[port_orig.find(":") + 1 :]
-        if desc == "":
-            desc = "device"
-        self._port.set(port)
-        self._port_desc.set(desc)
 
     def _on_ubx_config(self, *args, **kwargs):  # pylint: disable=unused-argument
         """
@@ -478,59 +356,11 @@ class SettingsFrame(Frame):
             self.__app.set_status("")
             self.__app.serial_handler.connect_file()
 
-    def _toggle_advanced(self):
-        """
-        Toggle advanced serial port settings panel on or off
-        """
-
-        self._show_advanced = not self._show_advanced
-        if self._show_advanced:
-            self._frm_advanced.grid(column=0, row=1, columnspan=3, sticky=(W, E))
-            self._btn_toggle.config(text=ADVON)
-        else:
-            self._frm_advanced.grid_forget()
-            self._btn_toggle.config(text=ADVOFF)
-
-    def _get_ports(self):
-        """
-        Populate list of available serial ports using pyserial comports tool.
-        If no ports found, disable all connection-dependent widgets.
-
-        Attempt to preselect the first port that has a recognisable
-        GPS designation in its description (usually only works on
-        Posix platforms - Windows doesn't parse UART device desc or HWID)
-        """
-
-        self._ports = sorted(comports())
-        init_idx = 0
-        port = ""
-        desc = ""
-        if len(self._ports) > 0:
-            for idx, (port, desc, _) in enumerate(self._ports, 1):
-                self._lbx_port.insert(idx, port + ": " + desc)
-                for kgp in KNOWNGPS:
-                    if kgp in desc:
-                        init_idx = idx
-                        break
-            self._noports = False
-        else:
-            self._noports = True
-            self.set_controls(NOPORTS)
-        self._lbx_port.activate(init_idx)
-        self._port.set(port)
-        self._port_desc.set(desc)
-
     def _reset(self):
         """
         Reset settings to defaults.
         """
 
-        self._baudrate.set(BAUDRATES[4])  # 9600
-        self._databits.set(8)
-        self._stopbits.set(1)
-        self._parity.set("None")
-        self._rtscts.set(False)
-        self._xonxoff.set(False)
         self._protocol.set(MIXED_PROTOCOL)
         self._format.set(DDD)
         self._units.set(UMM)
@@ -553,34 +383,8 @@ class SettingsFrame(Frame):
         :param int status: connection status as integer (0,1,2)
         """
 
-        self._lbl_port.configure(state=(NORMAL if status == DISCONNECTED else DISABLED))
-        self._lbx_port.configure(state=(NORMAL if status == DISCONNECTED else DISABLED))
-        self._lbl_baudrate.configure(
-            state=(NORMAL if status == DISCONNECTED else DISABLED)
-        )
-        self._spn_baudrate.configure(
-            state=(READONLY if status == DISCONNECTED else DISABLED)
-        )
-        self._lbl_databits.configure(
-            state=(NORMAL if status == DISCONNECTED else DISABLED)
-        )
-        self._spn_databits.configure(
-            state=(READONLY if status == DISCONNECTED else DISABLED)
-        )
-        self._lbl_stopbits.configure(
-            state=(NORMAL if status == DISCONNECTED else DISABLED)
-        )
-        self._spn_stopbits.configure(
-            state=(READONLY if status == DISCONNECTED else DISABLED)
-        )
-        self._lbl_parity.configure(
-            state=(NORMAL if status == DISCONNECTED else DISABLED)
-        )
-        self._spn_parity.configure(
-            state=(READONLY if status == DISCONNECTED else DISABLED)
-        )
-        self._chk_rts.configure(state=(NORMAL if status == DISCONNECTED else DISABLED))
-        self._chk_xon.configure(state=(NORMAL if status == DISCONNECTED else DISABLED))
+        self._frm_serial.set_controls(status)
+
         self._btn_connect.config(
             state=(
                 DISABLED if status in (CONNECTED, CONNECTED_FILE, NOPORTS) else NORMAL
@@ -624,15 +428,16 @@ class SettingsFrame(Frame):
         :rtype dict
         """
 
-        self._settings["port"] = self._port.get()
-        self._settings["noports"] = self._noports
-        self._settings["port_desc"] = self._port_desc.get()
-        self._settings["baudrate"] = self._baudrate.get()
-        self._settings["databits"] = self._databits.get()
-        self._settings["stopbits"] = self._stopbits.get()
-        self._settings["parity"] = self._parity.get()
-        self._settings["rtscts"] = self._rtscts.get()
-        self._settings["xonxoff"] = self._xonxoff.get()
+        self._settings["noports"] = self._frm_serial.noports
+        self._settings["port"] = self._frm_serial.port
+        self._settings["port_desc"] = self._frm_serial.port_desc
+        self._settings["baudrate"] = self._frm_serial.baudrate
+        self._settings["databits"] = self._frm_serial.databits
+        self._settings["stopbits"] = self._frm_serial.stopbits
+        self._settings["parity"] = self._frm_serial.parity
+        self._settings["rtscts"] = self._frm_serial.rtscts
+        self._settings["xonxoff"] = self._frm_serial.xonxoff
+
         self._settings["protocol"] = self._protocol.get()
         self._settings["raw"] = self._raw.get()
         self._settings["autoscroll"] = self._autoscroll.get()